DrmHwc2 : All buffer usage amend int to uint64_t
In order to adapter Gralloc 4.0.

Version : 1.3.86

+++ include/drmbuffer.h
@@ -42,7 +42,7 @@ public:
   int GetStride();
   int GetByteStride();
   int GetSize();
-  int GetUsage();
+  uint64_t GetUsage();
   int SetCrop(int left, int top, int right, int bottom);
   int GetCrop(int *left, int *top, int *right, int *bottom);
   uint32_t GetFourccFormat();
