+++ include/rockchip/drmgralloc.h
@@ -101,7 +101,7 @@ public:
   int hwc_get_handle_stride(buffer_handle_t hnd);
   int hwc_get_handle_byte_stride(buffer_handle_t hnd);
   int hwc_get_handle_byte_stride_workround(buffer_handle_t hnd);
-  int hwc_get_handle_usage(buffer_handle_t hnd);
+  uint64_t hwc_get_handle_usage(buffer_handle_t hnd);
   int hwc_get_handle_size(buffer_handle_t hnd);
   int hwc_get_handle_attributes(buffer_handle_t hnd, std::vector<int> *attrs);
   int hwc_get_handle_attibute(buffer_handle_t hnd, attribute_flag_t flag);
